feat(deposit): update data type and add integration testing

+++ src/main/java/com/wif3006/banking_system/deposit/dto/DepositFundsDto.java
@@ -1,10 +1,12 @@
 package com.wif3006.banking_system.deposit.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class DepositFundsDto {
     private String identificationNo;
     private String password;
-    private int amount;
+    private BigDecimal amount;
 }
